Autonomous Email Agent/requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
google-generativeai
python-multipart
emails
pandas
openpyxl
